project-template/frontend/pyproject.toml
[tool.poetry]
name = "frontend"
version = "0.1.0"
description = ""
authors = ["haruiz <henryruiz22@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.38.0"
watchdog = "2.1.5"
pandas = "^2.2.2"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
